refactor: organize transaction domain models
