+++ src/services/portfolio_transaction_processing_service/app/domain/position/history.py
@@ -12,7 +12,7 @@ from portfolio_common.ca_bundle_a_ordering import (
     ca_bundle_a_target_order_key,
 )
 
-from ..booked_transaction import BookedTransaction
+from ..transaction.booked import BookedTransaction
 from .reducer import PositionBalanceState, calculate_next_position_state
 
 PositionTransactionOrderKey = tuple[date, datetime, int, int, str, datetime, str]
