+++ src/services/portfolio_transaction_processing_service/app/domain/position/reducer.py
@@ -7,7 +7,7 @@ from datetime import date, timedelta
 from decimal import Decimal
 from typing import Callable
 
-from ..booked_transaction import BookedTransaction
+from ..transaction.booked import BookedTransaction
 
 CASH_POSITION_INFLOW_TRANSACTION_TYPES = {"DEPOSIT"}
 CASH_POSITION_OUTFLOW_TRANSACTION_TYPES = {"WITHDRAWAL", "FEE", "TAX"}
