+++ src/services/portfolio_transaction_processing_service/app/domain/transaction/__init__.py
@@ -1,0 +1,19 @@
+"""Canonical booked transactions and deterministic transaction identity."""
+
+from .booked import BookedTransaction
+from .semantic_identity import (
+    TRANSACTION_CORRECTION_IDENTITY_VERSION,
+    TRANSACTION_SEMANTIC_IDENTITY_VERSION,
+    TransactionSemanticIdentity,
+    build_transaction_correction_identity,
+    build_transaction_semantic_identity,
+)
+
+__all__ = [
+    "BookedTransaction",
+    "TRANSACTION_CORRECTION_IDENTITY_VERSION",
+    "TRANSACTION_SEMANTIC_IDENTITY_VERSION",
+    "TransactionSemanticIdentity",
+    "build_transaction_correction_identity",
+    "build_transaction_semantic_identity",
+]
