+++ src/services/portfolio_transaction_processing_service/app/domain/transaction/booked.py
@@ -1,3 +1,5 @@
+"""Define the canonical framework-neutral booked transaction command."""
+
 from __future__ import annotations
 
 from dataclasses import dataclass
